Refactor Tile into a separate React component
It's OK to have child React components within parent React components;
I'm not sure why previously I was getting "Objects are not valid as a
React child".

Perhaps because I was trying to render an Object directly to the DOM,
whereas now (since I'm now a bit more familiar with JSX syntax) I'm
assigning the object directly as props.